requirements.txt
Django==4.0
psycopg2_binary==2.9.2
djangorestframework==3.14.0
django-filter == 23.2
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
Pillow==10.0.0
pytest-django==4.5.2